Feature: Allow multiple notifications for action
Using the new Incoming Webhooks in `Slack` API, you can only
send notification to a single channel (defined in the
webhook). Previously, if you wanted to send to multiple
channels you could've specified multiple channels, but that
won't work with the new webhook system.

The solution for this is allowing multiple notifications
on the same action, and sending the same message to multiple
webhooks.

= Slack Bot =
Refactor #send_message to always get an array of urls from
 #get_webhook_urls and use them all.
Refactor #get_webhook_urls to check if `webhook_id` is an
array or a single id and always return an array of urls

= Notification Type =
Set each action in #notif_webhooks to be an array itself
and concatenate all the webhook_ids